requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
requests
networkx
streamlit
pandas
numpy
matplotlib
reportlab
transformers
torch
torchaudio
soundfile
Pillow
tqdm
torchvision
